Update class-plugin-name.php
I ran into an issue using the boilerplate when I needed to extend a class (I was using an autloader too, though I don't know if that makes a difference).

When trying to instantiate an extended class, the way the boilerplate is currently written, it seemed to create an object of the parent class. After spending some time on StackOverflow, it appeared that a slight alteration to your code to use the static method would solve the issue.

I updated the get_instance() method to reflect the changes.

P.S. sorry if some of my lingo is incorrect but hope you get the gist of what I'm saying!