Added the Blipfoto API
Modified the Blipfoto namespaces from those in the versions obtained
from the GitHub repo so that they’re unique to this plugin and widget
(I hope).
Added a class to act as an interface between the widget and the
Polaroid|Blipfoto API.
Moved the requirement of the settings from blipper-admin to
blipped-widget.  This may be moved again in the future.

Also tidied up the settings file.  Next thing to do with this is to get
the settings stored as arrays instead of discrete parameters.
Also added a tiny amount of documentation to class-wp-blipper.php.